test(ws): move maxPayload tests to a dedicated file
The maxPayload tests are in-process, but ws.test.ts also has client
tests that spawn a fresh debug subprocess per test under a 1000ms
deadline; those time out under the ASAN debug build and are unrelated
to this change. Move the maxPayload suite to its own file so it can be
verified in isolation.
